+++ cmd/src/batch_apply.go
@@ -69,6 +69,7 @@ Examples:
 
 		svc := batches.NewService(&batches.ServiceOpts{
 			AllowUnsupported: flags.allowUnsupported,
+			AllowIgnored:     flags.allowIgnored,
 			Client:           cfg.apiClient(flags.api, flagSet.Output()),
 			Workspace:        flags.workspace,
 		})