+++ cmd/src/batch_common.go
@@ -31,6 +31,7 @@ var (
 
 type batchApplyFlags struct {
 	allowUnsupported bool
+	allowIgnored     bool
 	api              *api.Flags
 	apply            bool
 	cacheDir         string
@@ -54,6 +55,10 @@ func newBatchApplyFlags(flagSet *flag.FlagSet, cacheDir, tempDir string) *batchA
 	flagSet.BoolVar(
 		&caf.allowUnsupported, "allow-unsupported", false,
 		"Allow unsupported code hosts.",
+	)
+	flagSet.BoolVar(
+		&caf.allowIgnored, "force-override-ignore", false,
+		"Do not ignore repositories that have a .batchignore file.",
 	)
 	flagSet.BoolVar(
 		&caf.apply, "apply", false,
@@ -234,6 +239,14 @@ func batchExecute(ctx context.Context, out *output.Output, svc *batches.Service,
 			batchCompletePending(pending, "Resolved repositories")
 
 			block := out.Block(output.Line(" ", output.StyleWarning, "Some repositories are hosted on unsupported code hosts and will be skipped. Use the -allow-unsupported flag to avoid skipping them."))
+			for repo := range repoSet {
+				block.Write(repo.Name)
+			}
+			block.Close()
+		} else if repoSet, ok := err.(batches.IgnoredRepoSet); ok {
+			batchCompletePending(pending, "Resolved repositories")
+
+			block := out.Block(output.Line(" ", output.StyleWarning, "The repositories listed below contain .batchignore files and will be skipped. Use the -force-override-ignore flag to avoid skipping them."))
 			for repo := range repoSet {
 				block.Write(repo.Name)
 			}